refactor: Remove rmi::Interface trait
This commit contains the refactoring commit related to the earlier
discussion (#231). Previously, RMI interface trait was mainly used
to sidestep the dependency issue between monitor and armv9a. Now
that the dependency issue has been resolved by the integration
(#175), the trait itslef doesn't have to be maintained.

[before]
rmm/src/realm/registry.rs contained most of the detailed RMI
implementation with trait.

[after]
The below contains each module's corresponding RMI implementation or
auxiliary functions without trait.
{
  rmm/src/realm/vcpu.rs       { create_vcpu }
  rmm/src/realm/timer.rs      { send_(timer_)state_to_host }
  rmm/src/realm/context.rs    { set_reg, get_reg }
  rmm/src/rtt.rs              { (rtt_)create, (rtt_)destroy, ... }
  rmm/src/gic.rs              { send_(gic_)state_to_host, receive_(gic_)state_from_host }
  rmm/src/mmio.rs             { emulate_mmio }
  ...
}
