+++ rmm/src/realm/timer.rs
@@ -1,5 +1,9 @@
 use super::context::Context;
+use crate::realm::registry::get_realm;
 use crate::realm::vcpu::VCPU;
+use crate::rmi::error::Error;
+use crate::rmi::error::InternalError::*;
+use crate::rmi::rec::run::Run;
 
 use armv9a::regs::*;
 
@@ -35,4 +39,22 @@ pub fn save_state(vcpu: &mut VCPU<Context>) {
     *&mut timer.cntp_cval_el0 = unsafe { CNTP_CVAL_EL0.get() };
     *&mut timer.cntp_ctl_el0 = unsafe { CNTP_CTL_EL0.get() };
     *&mut timer.cnthctl_el2 = unsafe { S3_4_C14_C1_0.get() };
+}
+
+pub fn send_state_to_host(id: usize, vcpu: usize, run: &mut Run) -> Result<(), Error> {
+    let realm = get_realm(id).ok_or(Error::RmiErrorOthers(NotExistRealm))?;
+    let mut locked_realm = realm.lock();
+    let vcpu = locked_realm
+        .vcpus
+        .get_mut(vcpu)
+        .ok_or(Error::RmiErrorOthers(NotExistVCPU))?;
+    let timer = &vcpu.lock().context.timer;
+
+    unsafe {
+        run.set_cntv_ctl(timer.cntv_ctl_el0);
+        run.set_cntv_cval(timer.cntv_cval_el0 - timer.cntvoff_el2);
+        run.set_cntp_ctl(timer.cntp_ctl_el0);
+        run.set_cntp_cval(timer.cntp_cval_el0 - timer.cntpoff_el2);
+    }
+    Ok(())
 }
