+++ rmm/src/rsi/mod.rs
@@ -11,6 +11,8 @@ use crate::listen;
 use crate::measurement::{
     HashContext, Measurement, MeasurementError, MEASUREMENTS_SLOT_NR, MEASUREMENTS_SLOT_RIM,
 };
+use crate::realm::config::realm_config;
+use crate::realm::context::{get_reg, set_reg};
 use crate::realm::mm::address::GuestPhysAddr;
 use crate::realm::mm::stage2_tte::invalid_ripas;
 use crate::rmi;
@@ -48,15 +50,14 @@ extern crate alloc;
 pub fn do_host_call(
     _arg: &[usize],
     ret: &mut [usize],
-    rmm: &Monitor,
+    _rmm: &Monitor,
     rec: &mut Rec<'_>,
     run: &mut Run,
 ) -> core::result::Result<(), Error> {
-    let rmi = rmm.rmi;
     let vcpuid = rec.vcpuid();
     let realmid = rec.realmid()?;
 
-    let ipa = rmi.get_reg(realmid, vcpuid, 1).unwrap_or(0x0);
+    let ipa = get_reg(realmid, vcpuid, 1).unwrap_or(0x0);
 
     let pa = crate::realm::registry::get_realm(realmid)
         .ok_or(Error::RmiErrorOthers(NotExistRealm))?
@@ -112,15 +113,14 @@ pub trait Interface {
 }
 
 pub fn set_event_handler(rsi: &mut RsiHandle) {
-    listen!(rsi, ATTEST_TOKEN_INIT, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
+    listen!(rsi, ATTEST_TOKEN_INIT, |_arg, ret, _rmm, rec, _| {
         let realmid = rec.realmid()?;
         let vcpuid = rec.vcpuid();
 
         let mut challenge: [u8; 64] = [0; 64];
 
         for i in 0..8 {
-            let challenge_part = rmi.get_reg(realmid, vcpuid, i + 2)?;
+            let challenge_part = get_reg(realmid, vcpuid, i + 2)?;
             let start_idx = i * 8;
             let end_idx = start_idx + 8;
             challenge[start_idx..end_idx].copy_from_slice(&challenge_part.to_le_bytes());
@@ -129,17 +129,16 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         rec.set_attest_challenge(&challenge);
         rec.set_attest_state(RmmRecAttestState::AttestInProgress);
 
-        rmi.set_reg(realmid, vcpuid, 0, SUCCESS)?;
+        set_reg(realmid, vcpuid, 0, SUCCESS)?;
 
         // TODO: Calculate real token size
-        rmi.set_reg(realmid, vcpuid, 1, 4096)?;
+        set_reg(realmid, vcpuid, 1, 4096)?;
 
         ret[0] = rmi::SUCCESS_REC_ENTER;
         Ok(())
     });
 
     listen!(rsi, ATTEST_TOKEN_CONTINUE, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
         let realmid = rec.realmid()?;
         let ipa_bits = rec.ipa_bits()?;
 
@@ -151,15 +150,15 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
 
         if rec.attest_state() != RmmRecAttestState::AttestInProgress {
             warn!("Calling attest token continue without init");
-            rmi.set_reg(realmid, vcpuid, 0, ERROR_STATE)?;
+            set_reg(realmid, vcpuid, 0, ERROR_STATE)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
             return Ok(());
         }
 
-        let attest_ipa = rmi.get_reg(realmid, vcpuid, 1)?;
+        let attest_ipa = get_reg(realmid, vcpuid, 1)?;
         if validate_ipa(attest_ipa, ipa_bits).is_err() {
             warn!("Wrong ipa passed {}", attest_ipa);
-            rmi.set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
+            set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
             return Ok(());
         }
@@ -184,8 +183,8 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
             hash_algo,
         );
 
-        rmi.set_reg(realmid, vcpuid, 0, SUCCESS)?;
-        rmi.set_reg(realmid, vcpuid, 1, attest_size)?;
+        set_reg(realmid, vcpuid, 0, SUCCESS)?;
+        set_reg(realmid, vcpuid, 1, attest_size)?;
 
         ret[0] = rmi::SUCCESS_REC_ENTER;
         Ok(())
@@ -193,12 +192,11 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
 
     listen!(rsi, HOST_CALL, do_host_call);
 
-    listen!(rsi, ABI_VERSION, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
+    listen!(rsi, ABI_VERSION, |_arg, ret, _rmm, rec, _| {
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
 
-        if rmi.set_reg(realmid, vcpuid, 0, VERSION).is_err() {
+        if set_reg(realmid, vcpuid, 0, VERSION).is_err() {
             warn!(
                 "Unable to set register 0. realmid: {:?} vcpuid: {:?}",
                 realmid, vcpuid
@@ -210,28 +208,27 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
     });
 
     listen!(rsi, MEASUREMENT_READ, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
         let mut measurement = Measurement::empty();
-        let index = rmi.get_reg(realmid, vcpuid, 1)?;
+        let index = get_reg(realmid, vcpuid, 1)?;
 
         if index >= MEASUREMENTS_SLOT_NR {
             warn!("Wrong index passed: {}", index);
-            rmi.set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
+            set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
             return Ok(());
         }
 
         rmm.rsi.measurement_read(realmid, index, &mut measurement)?;
-        rmi.set_reg(realmid, vcpuid, 0, SUCCESS)?;
+        set_reg(realmid, vcpuid, 0, SUCCESS)?;
         for (ind, chunk) in measurement
             .as_slice()
             .chunks_exact(core::mem::size_of::<usize>())
             .enumerate()
         {
             let reg_value = usize::from_le_bytes(chunk.try_into().unwrap());
-            rmi.set_reg(realmid, vcpuid, ind + 1, reg_value)?;
+            set_reg(realmid, vcpuid, ind + 1, reg_value)?;
         }
 
         ret[0] = rmi::SUCCESS_REC_ENTER;
@@ -239,20 +236,16 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
     });
 
     listen!(rsi, MEASUREMENT_EXTEND, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
 
-        let index = rmi.get_reg(realmid, vcpuid, 1)?;
-        let size = rmi.get_reg(realmid, vcpuid, 2)?;
+        let index = get_reg(realmid, vcpuid, 1)?;
+        let size = get_reg(realmid, vcpuid, 2)?;
         let mut buffer = [0u8; 64];
 
         for i in 0..8 {
-            buffer[i * 8..i * 8 + 8].copy_from_slice(
-                rmi.get_reg(realmid, vcpuid, i + 3)?
-                    .to_le_bytes()
-                    .as_slice(),
-            );
+            buffer[i * 8..i * 8 + 8]
+                .copy_from_slice(get_reg(realmid, vcpuid, i + 3)?.to_le_bytes().as_slice());
         }
 
         if size > buffer.len() || index == MEASUREMENTS_SLOT_RIM || index >= MEASUREMENTS_SLOT_NR {
@@ -260,7 +253,7 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
                 "Wrong index or buffer size passed: idx: {}, size: {}",
                 index, size
             );
-            rmi.set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
+            set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
             return Ok(());
         }
@@ -269,26 +262,25 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         let rd = rd.content::<Rd>();
         HashContext::new(&rmm.rsi, &rd)?.extend_measurement(&buffer[0..size], index)?;
 
-        rmi.set_reg(realmid, vcpuid, 0, SUCCESS)?;
+        set_reg(realmid, vcpuid, 0, SUCCESS)?;
         ret[0] = rmi::SUCCESS_REC_ENTER;
         Ok(())
     });
 
-    listen!(rsi, REALM_CONFIG, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
+    listen!(rsi, REALM_CONFIG, |_arg, ret, _rmm, rec, _| {
         let vcpuid = rec.vcpuid();
         let ipa_bits = rec.ipa_bits()?;
         let realmid = rec.realmid()?;
-        let config_ipa = rmi.get_reg(realmid, vcpuid, 1)?;
+        let config_ipa = get_reg(realmid, vcpuid, 1)?;
         if validate_ipa(config_ipa, ipa_bits).is_err() {
-            rmi.set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
+            set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
             return Ok(());
         }
 
-        rmi.realm_config(realmid, config_ipa, ipa_bits)?;
+        realm_config(realmid, config_ipa, ipa_bits)?;
 
-        if rmi.set_reg(realmid, vcpuid, 0, SUCCESS).is_err() {
+        if set_reg(realmid, vcpuid, 0, SUCCESS).is_err() {
             warn!(
                 "Unable to set register 0. realmid: {:?} vcpuid: {:?}",
                 realmid, vcpuid
@@ -298,15 +290,14 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         Ok(())
     });
 
-    listen!(rsi, IPA_STATE_GET, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
+    listen!(rsi, IPA_STATE_GET, |_arg, ret, _rmm, rec, _| {
         let vcpuid = rec.vcpuid();
         let ipa_bits = rec.ipa_bits()?;
         let realmid = rec.realmid()?;
 
-        let ipa_page = rmi.get_reg(realmid, vcpuid, 1)?;
+        let ipa_page = get_reg(realmid, vcpuid, 1)?;
         if validate_ipa(ipa_page, ipa_bits).is_err() {
-            if rmi.set_reg(realmid, vcpuid, 0, ERROR_INPUT).is_err() {
+            if set_reg(realmid, vcpuid, 0, ERROR_INPUT).is_err() {
                 warn!(
                     "Unable to set register 0. realmid: {:?} vcpuid: {:?}",
                     realmid, vcpuid
@@ -316,21 +307,21 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
             return Ok(());
         }
 
-        let ripas = rmi.rtt_get_ripas(realmid, ipa_page, RTT_PAGE_LEVEL)? as usize;
+        let ripas = crate::rtt::get_ripas(realmid, ipa_page, RTT_PAGE_LEVEL)? as usize;
 
         debug!(
             "RSI_IPA_STATE_GET: ipa_page: {:X} ripas: {:X}",
             ipa_page, ripas
         );
 
-        if rmi.set_reg(realmid, vcpuid, 0, SUCCESS).is_err() {
+        if set_reg(realmid, vcpuid, 0, SUCCESS).is_err() {
             warn!(
                 "Unable to set register 0. realmid: {:?} vcpuid: {:?}",
                 realmid, vcpuid
             );
         }
 
-        if rmi.set_reg(realmid, vcpuid, 1, ripas).is_err() {
+        if set_reg(realmid, vcpuid, 1, ripas).is_err() {
             warn!(
                 "Unable to set register 1. realmid: {:?} vcpuid: {:?}",
                 realmid, vcpuid
@@ -341,15 +332,14 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         Ok(())
     });
 
-    listen!(rsi, IPA_STATE_SET, |_arg, ret, rmm, rec, run| {
-        let rmi = rmm.rmi;
+    listen!(rsi, IPA_STATE_SET, |_arg, ret, _rmm, rec, run| {
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
         let ipa_bits = rec.ipa_bits()?;
 
-        let ipa_start = rmi.get_reg(realmid, vcpuid, 1)?;
-        let ipa_size = rmi.get_reg(realmid, vcpuid, 2)?;
-        let ipa_state = rmi.get_reg(realmid, vcpuid, 3)? as u8;
+        let ipa_start = get_reg(realmid, vcpuid, 1)?;
+        let ipa_size = get_reg(realmid, vcpuid, 2)?;
+        let ipa_state = get_reg(realmid, vcpuid, 3)? as u8;
         let ipa_end = ipa_start + ipa_size;
 
         if ipa_end <= ipa_start {
@@ -362,7 +352,7 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
             || !is_protected_ipa(ipa_start, ipa_bits)
             || !is_protected_ipa(ipa_end - 1, ipa_bits)
         {
-            rmi.set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
+            set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
             return Ok(());
         }
